requirements.txt
requests==2.32.3
psycopg2==2.9.9
python-dotenv==1.0.1
pytz==2024.2